feat: refactor API structure and remove unused files; consolidate experience and status APIs
