+++ src/manman/host/api/shared/__init__.py
@@ -1,0 +1,10 @@
+from .injectors import get_access_token, has_basic_worker_authz, inject_rmq_channel
+from .models import CurrentInstanceResponse, StdinCommandRequest
+
+__all__ = [
+    "inject_rmq_channel",
+    "get_access_token",
+    "has_basic_worker_authz",
+    "StdinCommandRequest",
+    "CurrentInstanceResponse",
+]
