+++ src/manman/host/api/shared/injectors.py
