+++ src/manman/host/api/status/__init__.py
@@ -1,0 +1,3 @@
+from .api import router
+
+__all__ = ["router"]
