+++ src/manman/host/api/worker_dal/__init__.py
@@ -1,0 +1,4 @@
+from .server import router as server_router
+from .worker import router as worker_router
+
+__all__ = ["worker_router", "server_router"]
